+++ src/Typography/Body/README.md
@@ -3,7 +3,7 @@ Ordinary body text.
 ### Usage
 
 ```jsx
-<Body>
+;<Body>
   Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed condimentum augue lorem. Suspendisse et tellus augue.
   Nulla lorem magna, efficitur eu condimentum molestie, varius et risus. Maecenas vitae arcu ligula. Fusce lobortis,
   eros consequat euismod pretium, risus lorem placerat leo, eu consectetur arcu tortor vitae ante. Proin malesuada dui
