+++ src/Typography/FinePrint/README.md
@@ -3,5 +3,5 @@ Very small text, a bit lighter, to add additionnal information.
 ### Usage
 
 ```jsx
-<FinePrint>I'm a fine print text</FinePrint>
+;<FinePrint>I'm a fine print text</FinePrint>
 ```
