+++ src/Typography/Title/README.md
@@ -3,5 +3,6 @@ A single page title. If you are using `<Page/>`, it is rendered internally anywa
 ### Usage
 
 ```jsx
-<Title>I'm a title</Title>
+import { Title } from "@operational/components"
+;<Title>I'm a title</Title>
 ```
